requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
pandas>=1.3.0
scikit-image>=0.18.0
joblib>=1.0.0
streamlit>=1.28.0
plotly>=5.0.0
streamlit-webrtc>=0.47.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
requests>=2.25.0
